+++ packages/components/src/composite/context.tsx
@@ -8,7 +8,9 @@ import { createContext, useContext } from '@wordpress/element';
  */
 import type { CompositeContextProps } from './types';
 
-export const CompositeContext = createContext< CompositeContextProps >( {} );
+export const CompositeContext = createContext<
+	CompositeContextProps | undefined
+>( undefined );
 CompositeContext.displayName = 'CompositeContext';
 
 export const useCompositeContext = () => useContext( CompositeContext );